+++ src/UKP/UKPmanager.cpp
@@ -50,8 +50,9 @@ void UKPmanager::begin()
     SSprintf("Kill switch enabled on pin:%d\n", KILL_SWITCH);
 #endif
     SSprintf("Number of buttons intialized:%d\n", button::buttonSum);
-    SSprintf("Number of profiles intialized:%d\n", defaultProfilesSum);
+    SSprintf("Number of profiles intialized:%d\n", maxProfiles);
 #endif
+    Retriever.begin();
 }
 
 #if PROFILES