+++ src/macrosengine/profiles.cpp
@@ -1,26 +1,30 @@
 #include "profiles.hpp"
 #if PROFILES
-
 profiles PR;
 int currentProfile = 0;
 
 void profiles::operator++(int)
 {
-    currentProfile < defaultProfilesSum - 1 ? currentProfile++ : currentProfile = 0;
+    currentProfile < maxProfiles - 1 ? currentProfile++ : currentProfile = 0;
 }
 
 void profiles::operator--(int)
 {
-    currentProfile < 0 ? currentProfile = defaultProfilesSum : currentProfile--;
+    currentProfile < 0 ? currentProfile = maxProfiles : currentProfile--;
 }
 
 void profiles::operator[](int i)
 {
-    if (i > defaultProfilesSum)
-        currentProfile = defaultProfilesSum;
+    if (i > maxProfiles)
+        currentProfile = maxProfiles;
     else if (i < 0)
         currentProfile = 0;
     else
         currentProfile = i;
+}
+
+void profiles::reset()
+{
+    currentProfile = 0;
 }
 #endif