+++ src/macrosengine/profiles.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "settings.h"
-extern const int defaultProfilesSum;
+extern int maxProfiles;
 #ifdef PROFILES
 
 class profiles
@@ -9,6 +9,7 @@ public:
     void operator++(int);
     void operator--(int);
     void operator[](int i);
+    void reset();
 };
 extern profiles Profile;
 extern int currentProfile;